+++ stable_baselines/deepq/dqn.py
@@ -7,7 +7,7 @@ from stable_baselines.common import tf_util, BaseRLModel, SetVerbosity, Tensorbo
 from stable_baselines.common.vec_env import VecEnv
 from stable_baselines.common.schedules import LinearSchedule
 from stable_baselines.deepq.replay_buffer import ReplayBuffer, PrioritizedReplayBuffer
-from stable_baselines.deepq.policies import DeepQPolicy
+from stable_baselines.deepq.policies import DQNPolicy
 from stable_baselines.a2c.utils import find_trainable_variables, total_episode_reward_logger
 
 
@@ -51,7 +51,7 @@ class DQN(BaseRLModel):
                  prioritized_replay_eps=1e-6, param_noise=False, verbose=0, tensorboard_log=None,
                  _init_setup_model=True):
 
-        super(DQN, self).__init__(policy=policy, env=env, verbose=verbose, policy_base=DeepQPolicy,
+        super(DQN, self).__init__(policy=policy, env=env, verbose=verbose, policy_base=DQNPolicy,
                                   requires_vec_env=False)
 
         self.checkpoint_path = checkpoint_path
@@ -93,8 +93,8 @@ class DQN(BaseRLModel):
 
             assert not isinstance(self.action_space, gym.spaces.Box), \
                 "Error: DQN cannot output a gym.spaces.Box action space."
-            assert issubclass(self.policy, DeepQPolicy), "Error: the input policy for the DQN model must be " \
-                                                         "an instance of DeepQPolicy."
+            assert issubclass(self.policy, DQNPolicy), "Error: the input policy for the DQN model must be " \
+                                                       "an instance of DQNPolicy."
 
             self.graph = tf.Graph()
             with self.graph.as_default():