+++ stable_baselines/deepq/experiments/enjoy_pong.py
@@ -10,7 +10,7 @@ def main():
     """
     env = gym.make("PongNoFrameskip-v4")
     env = deepq.wrap_atari_dqn(env)
-    model = DeepQ.load("pong_model.pkl", env)
+    model = DQN.load("pong_model.pkl", env)
 
     while True:
         obs, done = env.reset(), False